fix(naming): renamed and fixed bug
